bug fix uptime conversion

+++ kernel/sysinfo.h
@@ -2,7 +2,7 @@
 // Created by little-wolf on 3/26/23.
 //
 struct sysinfo{
-    long uptime;
+    double uptime;
     unsigned long totalram;
     unsigned long freeram;
     unsigned short procs;
